Hierarchical structure, remove containerd.
